move getSource to getSourceFromDisk
